cmd - rollup specific files
